refactor: :recycle: move types to directory, split internal and library types
